feat: Agentic AI apply code to golang standards
Motivation
- Apply the golang project standards to see how AI behaves
  - https://github.com/golang-standards/project-layout

Outcome
- The internal/ was created and code moved
  - pkg/ wasn't necessary to move
- The test were kept in cmd which is unexpected
- main.go was moved into cmd which is correcThe test were kept in cmd
  which is unexpected
- main.go was moved into cmd which is correct
