+++ cmd/gitversion-go/root.go
@@ -1,0 +1,22 @@
+package main
+
+import (
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "gitversion-go",
+	Short: "A Go implementation of GitVersion",
+	Long:  `A Go implementation of GitVersion to determine semantic versioning from a Git repository.`,
+}
+
+// Execute adds all child commands to the root command and sets flags appropriately.
+// This is called by main.main(). It only needs to happen once to the rootCmd.
+func Execute() {
+	err := rootCmd.Execute()
+	if err != nil {
+		os.Exit(1)
+	}
+}
