--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "gitversion-go/cmd"
-
-func main() {
-	cmd.Execute()
-}
